+++ test/test_helper_base.rb
@@ -40,11 +40,11 @@ class ActiveSupport::TestCase
 
   def setup
     super
-    Folio::Current.original_reset
+    Folio::Current.reset
   end
 
   def reset_folio_current(site_user_link)
-    ::Folio::Current.original_reset
+    ::Folio::Current.reset
     ::Folio::Current.nillify_site_records
     ::Folio::Current.user = site_user_link.user
 
@@ -63,7 +63,7 @@ class Cell::TestCase # do not inherit from ActiveSupport::TestCase
   attr_reader :site
 
   def setup
-    Folio::Current.original_reset
+    Folio::Current.reset
     @site = get_any_site
   end